feat : 에러 코드 추가 및 에러 처리 도메인 분리
